cleanup and reorganize
