--- src/client/methods/getServiceInstance.js
@@ -1,3 +1,0 @@
-export default function getServiceInstance () {
-
-}
