--- src/client/utils/index.js
@@ -1,3 +1,0 @@
-/**
- * Created by bhoriuchi on 12/1/16.
- */
